소영 가계부 - 컨트롤러에 넘어오는 값 12자리인 경우 고려해서 int -> long으로 변경(별로큰의미없음.)

+++ src/main/java/net/softsociety/secretary/controller/CashbookBudgetController.java
@@ -14,7 +14,7 @@ public class CashbookBudgetController {
 
 	/** 예산 입력 */
 	@PostMapping("setBudget")
-	public String setBudget(int budgetAmount) {
+	public String setBudget(long budgetAmount) {
 		log.debug("넘어온 예산:{}", budgetAmount);
 		
 		return "redirect:/cashbook/";
